Redact bsp website.
The website looked quite official thanks to the pretty pandoc theme I
found online. This  made it appear as if BSP is an approved standard
even if the introduction had a disclaimer that this was just a draft for
an idea.

For now, BSP is more of a thought experiment so I think it's better to
make the markdown file on GitHub the place to read about it.