Clarify `dependencySources` may return source files
Sources can be either files or directories. Bazel, for example,
does not know anything about source directories and every target
is mapped to a set of source files that need to be compiled.

For BSP to be compatible with build tools like Bazel, we need to make
this distinction clear.