Allow more fine-grained capabilities declaration
These changes to the spec allow us to distinguish between:

1. A build target that can be tested because it defines tests.
2. A server that cannot test a target because it doesn't support it.

On top of that, we remove the build target kinds from build target to
represent the capabilities of a target and instead go for a more direct
approach.

Olafur presented in #12 to have the capabilities explicit in the build
target. After some discussion with him and several iterations, we've
agreed to include `BuildTargetCapabilities`.

With these changes, a client can:

* Notify users that the test execution for a given language is not
  supported.
* Modify the UI to show that a target cannot be tested, or cannot be
  compiled.

Before this change, it was problematic that a server would support
compilation of, say, Kotlin but not test execution as there was no way
of telling cases 1) and 2) apart (above).

Fixes https://github.com/scalacenter/bsp/issues/12.