Add build target kind back
Even though `IntegrationTest` and `Bench` are not supported by, to the
best of my knowledge, any build tool and IDE, it is still beneficial to
have the kind in the build target.

For example, IntelliJ doesn't support `IntegrationTest`, but if IntelliJ
users want to see the tests that they can run in their projects,
IntelliJ has to show the test main class from both the `Test` target
kind and `IntegrationTest` target kind.

Another use case where this is useful is in showing all the imported
targets, no matter if there's support for them or not, in a list of the
UI so that the user can manually trigger a bsp action on them.