Implement bsp4j module based on org.eclipse.lsp4j.jsonrpc
Add a new module `ch.epfl.scala:bsp4j` that is written in pure Java in
the same style as the lsp4j module. The benefit of this module compared
to `ch.epfl.scala:bsp_2.12` is that it has a lower dependency footprint.

Includes a full end-to-end integration test with bloop that stresses the
current implemented endpoints in bloop.  The module implements the full
spec but we can't test all endpoints yet since no server implements
them.

Future work:
- add docstrings for the data structures and endpoints
- add `@NonNull` where necessary