Add alternative proposal for BSP Connection Protocol
I think I've made a breakthrough in the specification of the protocol
that improves substantially on the previous iteration of the connection
protocol.

The new iteration supports global BSP locations to write the connection
files and makes the protocol connection-agnostic, relieving BSP clients
with the need of doing complicated template variables or passing
information to the build tool commands to start the BSP servers.

This commit adds an exhaustive explanation, including examples, about
the new solution. It's now been designed with the following requirements
in mind:

1. Clients do not require beforehand knowledge about a specific server
   implementation to be able to connect to it.
2. Clients do not need to know the specifics of what kind of connections are
   supported by concrete build tools.
3. Clients can connect to build tools installed both at the machine and at the
   workspace level.
4. Multiple versions of the same build tool can have different connection
    mechanisms.
5. Multiple different BSP servers can run in the same workspace directory.