Add `reset` to `buildTarget/publishDiagnostics`
The current solution does not allow the client to handle the lifetime of
the diagnostics correctly. For example, when errors disappear in a
buffer, there is no way for the client to know that it should clean
them. Taking another approach in the lifetime handling brings us to
other problems like missing present warnings that are not present during
incremental compilation, see https://github.com/scalacenter/bloop/issues/696

After a discussion with Olafur, we have come up with the following
solution. Reset is a flag that allows servers to tell clients how they
should use the diagnostics. When `reset` is enabled, the behavior is the
same as in the Language Server Protocol. When it isn't enabled, it
allows the build tools to stream errors coming from, say, a compilation
run to the client and letting the client accumulate all of those until
the end of the compilation request.

The following commit makes changes to the code and the protocol spec to
reflect this new workflow.

Co-authored-by: Olafur Pall Geirsson <olafurpg@gmail.com>