Use `parents: Option[List[String]]` instead of parent
A build tool can run a task in a project because more than one project
depends on it. To avoid forcing the build tool to decide which should be
considered the parent, we instead add all the parents in the build tool
side and then let the client handle that data.

With this information, a client:

1. Can display task ids in a tree fashion by doing `.head` on the list.
2. Can display a task graph of all the dependencies to the user.