Remove streaming of diagnostics
In previous versions of BSP, we added a field `reset: Boolean` to
diagnostics that would allow the server to stream diagnostics to the
client, instead of sending them all at once.

The motivation for this feature is that the end user would get more
instant feedback than with the batched approach. The way this approach
worked is that the first diagnostic notification for a text document
identifier would have `reset: Boolean = true` and the rest would be set
to `false` preventing the client from clearing all diagnostics in the
same text document.

This commit instead removes this approach. This removal is motivated by
the real-world experience of using this diagnostics publishing strategy
in tools in the Scala ecosystem. There are two main reasons for its
removal:

1. Notifications in JSON-RPC have no ordering so they can arrive to the
client in whatever order they are sent/processed by the OS/client.
Therefore, the approach above could be not setting certain diagnostics
if they are sent before the diagnostic that sets `reset = true`. This
has happened to me several times while using these tools.

2. There are other ways we can provide instant feedback to the client.
Instead of streaming, the server can send small batches of diagnostics
for text document identifiers as the compiler goes through the code
(e.g. every time the compiler finishes a compilation phase).  This comes
at the cost of delivering more notifications to the client and minimizes
issues with race conditions when setting diagnostics.

Therefore, we stay with the LSP approach of sending diagnostics. It's
clearer and simpler for BSP clients.