Upgrade lsp4j dependency to v0.8.1 from v0.5.0.
The latest version seems to enforce the `@NotNull` annotation at
runtime. The annotation is verified with a `Preconditions.java` class
that is only included in the lsp4j module, which bsp4js doesn't depend
on, so the source code of `Preconditions.java` was copied verbatim into
this codebase along with an accompanying license.