Add `manual` build target tag (#133)
Add `manual` build target tag. This is based on the same manual tag in Bazel.

It is used when targets are not supposed to be considered by default when triggering a full workspace build. Instead a build/test/etc action on these targets should be triggered manually by the user.