Introduce workspace/reload request
The `reload` request instructs the build server to reload the build configuration.
This request should be supported by build tools that keep their state in memory.
If the `reload` request returns with an error, it's expected that other requests
respond with the previously known "good" state.

This request is motivated by tools such as sbt, which maintain an in-memory state of the build, where reloading the build on-demand is not considered practical.