Add in `noOp` field to `CompileReport`. (#196)
This is meant to close #148 by adding in another field to the
`CopileReport` to signify to the client that the compilation was a
no-op. This is useful for some clients that trigger things based on
compliations, and they want to skip that trigger if the compilation was
a no-op. Currently this is accomplished by looking for keywords in the
message as is the case with Metals and Bloop.