Update/org.eclipse.lsp4j.generator 0.19.0 (#429)
ResponseErrorCode.jsonrpcReservedErrorRangeEnd has never been a valid code
previously it was named "serverErrorEnd", but it didn't mean "cannot handle requests after shutdown",
it was a marker. From the docs: "This is the start range of JSON RPC reserved error codes.
It doesn't denote a real error code."

I'm not changing it now, because it's a breaking change.