chore(governance): add in some basic things for maintenance
This is a PR that came about during the BSP discussions at the Scala
Tooling Summit. This change updates the README and the CONTRIBUTING doc
with some more relevant information about how this project is
maintained and also added some guidelines. It also adds a CODEOWNERS and
a feature request template.