docs: align BSP diagnostics clearing with LSP behavior
Update the BSP specification to clarify that servers should send
`build/publishDiagnostics` with an empty diagnostics array to clear
former diagnostics.
This change addresses the inconsistencies among various BSP servers
and aligns the behavior with the Language Server Protoco,
providing a more consistent experience for clients such as `scalameta/metals`.

https://microsoft.github.io/language-server-protocol/specifications/lsp/3.17/specification/#textDocument_publishDiagnostics

https://github.com/build-server-protocol/build-server-protocol/issues/481