Add functionality for reading stdin and printing stdout/stderr
This commit introduces new methods in BuildServer and BuildClient classes that allow for interaction with the current running target. A 'PrintParams' class is also added to encapsulate parameters for these methods. The added methods are 'onRunReadStdin' in BuildServer, and 'onRunPrintStdout' and 'onRunPrintStderr' in BuildClient. These changes allow passing user input to the target's stdin stream and receiving its stdout and stderr streams. The documentation and specifications have been updated accordingly.