Refactor running applications and tests using BSP
Expanded the functionality for running applications and tests using BSP. Added optional support for environment variables and working directory parameters. This allows the user to set up the specific environment for the tests or applications.

Arguments and environment variables lists in ScalaMainClass and ScalaTestSuites have been marked deprecated, because the base request now contains these arguments.