Add `Cancel` notification (#596)
In JVM languages, cancelling requests is done by using the
`CompletableFuture` feature, so the `Cancel` notification is
unnecessary. However, some languages (such as Rust) don't support this
feature. This means that servers/clients written in these technologies
should have this notification in their bsp library, so we would like it
to be explicitly defined in the protocol for proper generation process.

---------

Co-authored-by: Andrzej Głuszak <andrzej.gluszak@jetbrains.com>